+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Sprinkle list in the bakery `debug` command to uses the new `sprinkle:list` table
 - `routerCacheFile` config now only contains filename. Locator is used to find the full path
+- Moved some constants from `app/defines.php` to `app/sprinkles/core/defines.php`
 
 ### Fix
 - Fix for `Test` Bakery command
@@ -31,6 +32,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - Added specific tests for sprinkleManager with 100% test coverage
 - Ignore existing `package-lock.json` which caused incorrect dependencies to be installed when upgrading from older versions of UserFrosting.
 - Vendor assets not found in production mode
+
+### Deprecated
+- `UserFrosting\APP_DIR_NAME` : Use `UserFrosting\APP_DIR` instead.
 
 ## 4.2.0-beta.1